backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
neo4j>=5.15.0
pydantic>=2.7.0
python-dotenv>=1.0.0
fastmcp>=2.0.0
gitpython>=3.1.43
